requirements.txt
anyio==3.6.2
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.1
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
coverage==6.1.1
cryptography==37.0.2
cycler==0.11.0
dj-database-url==0.5.0
Django==3.2.8
django-adminlte2==0.4.3
django-bootstrap4==3.0.1
django-environ==0.10.0
django-filter==21.1
django-grappelli==2.15.3
django-stdimage==5.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
entrypoints==0.3
flake8==5.0.4
flake8-html==0.4.2
fonttools==4.33.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
Jinja2==3.1.2
kiwisolver==1.4.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.7.0
model-mommy==2.0.0
numpy==1.22.4
onesignal-sdk==2.0.0
outcome==1.1.0
packaging==21.3
Pillow==8.3.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.5.0
Pygments==2.13.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
whitenoise==5.3.0
wsproto==1.1.0
zipp==3.8.0
